Added staging OS images (formerly init) to prepare OS images for first time boot initialization. Added common staging. Move away from installing separate init scripts to generating the scripts on the fly in the make files. This avoids having to maintain separate files and the possibility of having to use sed to customize scripts for different mod configurations.
Added the octoprint initialization and a way to invoke it from the
first time script.

Now using /etc/rc.local for the first time script. This avoids the need
to run systemctl which doesn't work in a proot environment since
that is not an actual boot environment.

Completed the opiz staging which is now ready to begin boot testing. rpi3
needs to be updated to match.